Add Zluinav theme
Please note that this theme's screenshot might not be good enough. As I am a blind person, I can't sure if the screenshot looks good. I just wanted to submit this theme so that any blind people using this static site generator can use this theme.

As such, designs in this theme might not look good to sighted people!